restructures redux store
