+++ src/components/Welcome/Dashboard.js
@@ -2,11 +2,11 @@ import React from 'react';
 import Header from './Header';
 import DashboardBody from './DashboardBody';
 import { useDispatch, useSelector } from 'react-redux';
-import { setName } from '../../userSlice';
+import { setName } from '../../utils/redux/userSlice';
 
 export default function Dashboard() {
   const dispatch = useDispatch();
-  const { user } = useSelector((state) => state.thing);
+  const { user } = useSelector((state) => state.userData);
   return (
     <div
       style={{
