+++ src/index.js
@@ -3,14 +3,14 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import { BrowserRouter as Router } from 'react-router-dom';
-import { store } from './store';
-import { Provider } from 'react-redux';
+import { store } from './utils/redux/store';
+import { Provider as ReduxProvider } from 'react-redux';
 
 ReactDOM.render(
-  <Provider store={store}>
+  <ReduxProvider store={store}>
     <Router>
       <App />
     </Router>
-  </Provider>,
+  </ReduxProvider>,
   document.getElementById('root')
 );
